COMMERCE-1469 Move product indexers to internal package
