--- commerce-product-service/src/main/resources/com/liferay/commerce/product/search/packageinfo
@@ -1,1 +1,0 @@
-version 2.1.0
